ai-services/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
pydantic==2.6.1
python-dotenv==1.0.1
scikit-learn==1.4.1.post1
pandas==2.2.0
numpy==1.26.4
